pyproject.toml
[project]
name = "youtube-telegram-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Hesam Pourabbasian",email = "hesampourabbasian6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytube (>=15.0.0,<16.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "pytubefix (>=8.13.1,<9.0.0)",
    "yt-dlp (>=2025.4.30,<2026.0.0)"
]

[tool.poetry]
packages = [{include = "youtube_telegram_downloader", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
